Cast the 8 bit int to 16 before shifing

+++ firmware/Adafruit_SI1145.cpp
@@ -120,7 +120,7 @@ uint16_t Adafruit_SI1145::readVisible(void) {
  uint16_t val = 0;
 
  val = read8(SI1145_REG_ALSVISDATA0);
- val |= read8(SI1145_REG_ALSVISDATA1) << 8;
+ val |= (uint16_t)read8(SI1145_REG_ALSVISDATA1) << 8;
 
  return val;
 }
